Rename OpenCircuitParams and modules
